Fixed bugs to cast integral division result to double or float

+++ src/main/java/net/pms/dlna/DLNAResource.java
@@ -3351,7 +3351,7 @@ public abstract class DLNAResource extends HTTPResource implements Cloneable, Ru
 		}
 
 		if (low > 0 && media.getBitrate() > 0) {
-			lastStartPosition = (low * 8) / media.getBitrate();
+			lastStartPosition = (low * 8) / (double) media.getBitrate();
 			LOGGER.trace("Estimating seek position from byte range:");
 			LOGGER.trace("   media.getBitrate: " + media.getBitrate());
 			LOGGER.trace("   low: " + low);
@@ -3414,7 +3414,7 @@ public abstract class DLNAResource extends HTTPResource implements Cloneable, Ru
 					resume.update((Range.Time) range, this);
 				}
 
-				params.timeseek = resume.getTimeOffset() / 1000;
+				params.timeseek = resume.getTimeOffset() / (double) 1000;
 				if (player == null) {
 					player = new FFMpegVideo();
 				}
@@ -3590,7 +3590,7 @@ public abstract class DLNAResource extends HTTPResource implements Cloneable, Ru
 		) {
 			Double seekPosition = (double) configurationSpecificToRenderer.getThumbnailSeekPos();
 			if (isResume()) {
-				Double resumePosition = (double) (resume.getTimeOffset() / 1000);
+				Double resumePosition = resume.getTimeOffset() / (double) 1000;
 
 				if (media.getDurationInSeconds() > 0 && resumePosition < media.getDurationInSeconds()) {
 					seekPosition = resumePosition;
